+++ plugins/visualization/ovp-advanced-visualization/src/mTBoxAlgorithmStackedContinuousViz.hpp
@@ -59,7 +59,7 @@ namespace Mensia
 			explicit CBoxAlgorithmStackedContinuousVizListener(const std::vector<int>& vParameter)
 				: CBoxAlgorithmVizListener(vParameter) { }
 
-			bool onInputTypeChanged(OpenViBE::Kernel::IBox& box, const uint32_t index) override
+			bool onInputTypeChanged(OpenViBE::Kernel::IBox& box, const size_t index) override
 			{
 				OpenViBE::CIdentifier typeID = OV_UndefinedIdentifier;
 				box.getInputType(index, typeID);
