+++ plugins/visualization/ovp-advanced-visualization/src/mTBoxAlgorithmStackedInstantViz.hpp
@@ -65,7 +65,7 @@ namespace Mensia
 
 			CBoxAlgorithmStackedInstantVizListener(const std::vector<int>& vParameter) : CBoxAlgorithmVizListener(vParameter) { }
 
-			bool onInputTypeChanged(IBox& box, const uint32_t index) override
+			bool onInputTypeChanged(IBox& box, const size_t index) override
 			{
 				OpenViBE::CIdentifier typeID = OV_UndefinedIdentifier;
 				box.getInputType(index, typeID);
